error functions outsourcing to utils
