+++ examples/demo.html
@@ -49,7 +49,7 @@
             
                 <template class="live-example">
                     <h1>Example: Hello World!</h1>
-                    <script data-run="true" data-demo="true">
+                    <script data-run="true">
                         const helloWorld = () => {
                             console.log("Hello World!");
                         };
